fixade metoder

+++ src/DHSCA_v2/Temperature.java
@@ -1,9 +1,9 @@
 package DHSCA_v2;
 
 public class Temperature extends SensorValue {
-    private float degrees;
+    private double degrees;
 
-    public Temperature(float degrees, String timeStamp) {
+    public Temperature(double degrees, String timeStamp) {
         super(timeStamp);
         this.degrees = degrees;
     }
